backend/requirements.txt
fastapi~=0.115.8
pydantic-settings~=2.7.1
uvicorn~=0.34.0